fixing encodings only available images

+++ Server/Data/DataLogic.py
@@ -40,35 +40,38 @@ class suspectData():
         encodings_file={}
         for item in Items:
             filename = item['objectKey'].split("/")[1]
-            name = item['suspectName']
-            #print(filename, name)
-            try:
-                self.s3.Bucket(self.bucketName).download_file(item['objectKey'], f'./tmp/{filename}')
-            except botocore.exceptions.ClientError as e:
-                if e.response['Error']['Code'] == "404":
-                    logging.warning("No Image found in bucket")
-                else:
-                    raise
-            #print(encodings)        
-            suspect_face = face_recognition.load_image_file(f'./tmp/{filename}')
-            face_encodings = face_recognition.face_encodings(suspect_face) 
-            #print(face_encodings)
-            if face_encodings==[]:
-                logging.warning("Image is in wrong format")
-                #encodings_file[name] = ([])
+            if filename!="":
                 
-                #encodings.append(face_encodings) 
-                self.updateSuspect(name,face_encodings, "Bad Encodings")
-            else:
-                #saving info to array
-                encodings.append(face_encodings[0])
-                names.append(name)
-                #saving info to dictionary
-                encodings_file[name]= face_encodings[0]
-                # saving info to DB
-                encodings_string = face_encodings[0].tostring()
-                self.updateSuspect(name,encodings_string, "Good Encoding")
-            os.remove(f'./tmp/{filename}')
+            
+                name = item['suspectName']
+                #print(filename, name)
+                try:
+                    self.s3.Bucket(self.bucketName).download_file(item['objectKey'], f'./tmp/{filename}')
+                except botocore.exceptions.ClientError as e:
+                    if e.response['Error']['Code'] == "404":
+                        logging.warning("No Image found in bucket")
+                    else:
+                        raise
+                #print(encodings)        
+                suspect_face = face_recognition.load_image_file(f'./tmp/{filename}')
+                face_encodings = face_recognition.face_encodings(suspect_face) 
+                #print(face_encodings)
+                if face_encodings==[]:
+                    logging.warning("Image is in wrong format")
+                    #encodings_file[name] = ([])
+                    
+                    #encodings.append(face_encodings) 
+                    self.updateSuspect(name,face_encodings, "Bad Encodings")
+                else:
+                    #saving info to array
+                    encodings.append(face_encodings[0])
+                    names.append(name)
+                    #saving info to dictionary
+                    encodings_file[name]= face_encodings[0]
+                    # saving info to DB
+                    encodings_string = face_encodings[0].tostring()
+                    self.updateSuspect(name,encodings_string, "Good Encoding")
+                os.remove(f'./tmp/{filename}')
             
        
         dump(encodings_file, open(f"./tmp/{self.user}/encodings.dat","wb"))
